pyproject.toml
#
# Build system requirements
#
[build-system]

requires = ["setuptools", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

#
#
# Basic project metadata
#
[project]

name = "dragonET"

description = "Dodgy Reconstruction and Alignment of General OrientatioNs for Electron Tomography"

readme = "README.md"

requires-python = ">=3.6"

license = {file = "LICENSE"}

keywords = [
  "dragonET",
  "python",
  "project"
]

authors = [
  {name = "James Parkhurst", email = "james.parkhurst@diamond.ac.uk" }
]

classifiers = [
  "Development Status :: 3 - Alpha",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.6",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]

dependencies = [
  "matplotlib",
  "mrcfile",
  "napari[all]",
  "napari-mrcfile-reader",
  "numpy",
  "scipy",
  "pyyaml",
]

dynamic = [
  "version"
]

#
# Project urls
#
[project.urls]

"Source" = "https://github.com/jmp1985/dragonET"
"Tracker" = "https://github.com/jmp1985/dragonET/issues"

#
# Project additional optional dependencies
#
[project.optional-dependencies]

test = [
  "pytest",
  "pytest-cov",
]

#
# Project command line scripts
#
[project.scripts]

"dragonET.new" = "dragonET.command_line:new"
"dragonET.align" = "dragonET.command_line:align"
"dragonET.project" = "dragonET.command_line:project"
"dragonET.refine" = "dragonET.command_line:refine"
"dragonET.reconstruct" = "dragonET.command_line:reconstruct"
"dragonET.track" = "dragonET.command_line:track"
"dragonET.contours.extend" = "dragonET.command_line:contours_extend"
"dragonET.contours.pick" = "dragonET.command_line:contours_pick"
"dragonET.contours.refine" = "dragonET.command_line:contours_refine"
"dragonET.contours.triangulate" = "dragonET.command_line:contours_triangulate"
"dragonET.stack.edit" = "dragonET.command_line:stack_edit"
"dragonET.stack.predict" = "dragonET.command_line:stack_predict"
"dragonET.stack.rebin" = "dragonET.command_line:stack_rebin"
"dragonET.stack.rot90" = "dragonET.command_line:stack_rot90"
"dragonET.stack.transform" = "dragonET.command_line:stack_transform"
"dragonET.volume.rebin" = "dragonET.command_line:volume_rebin"
"dragonET.volume.select_sample_axis" = "dragonET.command_line:volume_select_sample_axis"

#
# Configure setup tools
#
[tool.setuptools.packages.find]

where = ["src"]

#
# Set the package version from git tag
#
[tool.setuptools.dynamic]

version = {attr = "dragonET._version.__version__"}

#
# Configure setuptools-scm to generate version file
#
[tool.setuptools_scm]

write_to = "src/dragonET/_version.py"

#
# Configure pytest
#
[tool.pytest.ini_options]

addopts = "--cov=src --cov-report term --cov-report html"

testpaths = [ "tests" ]
